ujian_app/requirementsFix.txt
asgiref==3.8.1
audioread==3.0.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Django==4.2.20
django-bootstrap-v5==1.0.11
dlib==19.24.8
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
mysqlclient==2.2.7
networkx==3.4.2
numba==0.61.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
